Dependencies + Python support

+++ tasks/build.py
@@ -4,8 +4,10 @@ Build Tasks
 ~~~~~~~~~~~
 
 """
+from __future__ import print_function
 
 import os as _os
+import re as _re
 
 import invoke as _invoke
 
@@ -23,20 +25,78 @@ def source(ctx):
 
 
 @_invoke.task()
-def wheels(ctx):
+def wheels(ctx, arches=None):
     """ Build wheels """
+    path = 'wheel/dist'
+    if arches is None:
+        arches = "x86_64 i686"
+    arches = arches.split()
     with ctx.shell.root_dir():
-        ctx.shell.rm_rf('wheel/dist')
-        ctx.run(ctx.c('''
-            docker run --rm -it -v%s/wheel:/io
-            quay.io/pypa/manylinux1_x86_64:latest
-            /io/build.sh %s %s
-        ''', _os.getcwd(), ctx.package, "27 34 35 36 37"), pty=True)
-        ctx.run(ctx.c('''
-            docker run --rm -it -v%s/wheel:/io
-            quay.io/pypa/manylinux1_i686:latest
-            linux32 /io/build.sh %s %s
-        ''', _os.getcwd(), ctx.package, "27 34 35 36 37"), pty=True)
+        ctx.shell.rm_rf(path)
+        pythons = "36 37 38 39 310"
+
+        for arch in arches:
+            if arch in ("x86_64", "i686"):
+                ctx.run(
+                    ctx.c('''
+                        docker run --rm -it -v%s/wheel:/io
+                        quay.io/pypa/manylinux1_%s:latest
+                    ''' + (arch == "i686" and "linux32" or "") + '''
+                        /io/build.sh %s %s
+                    ''', _os.getcwd(), arch, ctx.package, "27"),
+                    echo=True, pty=True,
+                )
+            ctx.run(
+                ctx.c('''
+                    docker run --rm -it -v%s/wheel:/io
+                    quay.io/pypa/manylinux_2_24_%s:latest
+                    ''' + (arch == "i686" and "linux32" or "") + '''
+                    /io/build.sh %s %s
+                ''', _os.getcwd(), arch, ctx.package, pythons),
+                echo=True, pty=True,
+            )
+
+        # strip file names
+        multi_sub = _re.compile(r'(?:[.-]manylinux[^.]+)+').sub
+        tomove = []
+        for name in _os.listdir(path):
+            if not name.endswith('.whl'):
+                continue
+            pick = _best_manylinux(name)
+            if pick:
+                tomove.append((
+                    name, multi_sub((lambda _, p=pick: '-' + p), name)
+                ))
+        for old, new in tomove:
+            if old == new:
+                continue
+
+            print("%s -> %s" % (old, new))
+            _os.rename(_os.path.join(path, old), _os.path.join(path, new))
+
+
+def _best_manylinux(name):
+    """ Find best manylinux variant produced by auditwheel """
+    tag_list_search = _re.compile(r'(?:[.-]manylinux[^.]+)+').search
+
+    tags = tag_list_search(name)
+    if not tags:
+        return None
+    sortable = []
+    for tag in tags.group(0).strip('-.').split('.'):
+        assert tag.startswith('manylinux')
+        ver = tag[len('manylinux'):]
+        if ver.startswith(("1_", "2010_", "2014_")):
+            sortable.append(((1, int(ver.split('_')[0])), tag))
+            continue
+        assert ver.startswith('_')
+        ver = tuple(map(int, ver[1:].split('_')[:2]))
+        assert ver[0] > 1
+        sortable.append((ver, tag))
+    if not sortable:
+        return None
+
+    return min(sortable)[-1]
 
 
 @_invoke.task(_doc.doc)
